Disables the incomplete Scar feature (resolves #75)
Because the Scar mechanic was incomplete, it was causing players to take unexplained damage from an unidentified source, with no way to remove the effect except to die. I've commented out some of the code to disable the mechanic while leaving the code in place for any future development of the feature.